fix: improved validation architecture
