+++ packages/server/src/middleware/errorHandler.js
@@ -1,3 +1,5 @@
+import { createErrorResponse } from '../utils/responseHelpers.js';
+
 // Error handling middleware
 export const errorHandler = (err, req, res, _next) => {
   if (process.env.NODE_ENV === 'development') {
@@ -7,39 +9,60 @@ export const errorHandler = (err, req, res, _next) => {
   // Default error
   let statusCode = 500;
   let message = "Internal server error";
+  let code = null;
 
   // Handle different types of errors
   if (err.name === "ValidationError") {
     statusCode = 400;
     message = "Validation error";
+    code = "VALIDATION_ERROR";
   } else if (err.name === "UnauthorizedError") {
     statusCode = 401;
     message = "Unauthorized";
+    code = "UNAUTHORIZED_ERROR";
   } else if (err.name === "ForbiddenError") {
     statusCode = 403;
     message = "Forbidden";
+    code = "FORBIDDEN_ERROR";
   } else if (err.name === "NotFoundError") {
     statusCode = 404;
     message = "Resource not found";
+    code = "NOT_FOUND_ERROR";
   } else if (err.name === "ConflictError") {
     statusCode = 409;
     message = "Resource conflict";
+    code = "CONFLICT_ERROR";
   } else if (err.code === "23505") {
     // PostgreSQL unique constraint
     statusCode = 409;
     message = "Resource already exists";
+    code = "UNIQUE_CONSTRAINT_ERROR";
   } else if (err.code === "23503") {
     // PostgreSQL foreign key constraint
     statusCode = 400;
     message = "Invalid reference";
+    code = "FOREIGN_KEY_ERROR";
   }
 
-  // Send error response
-  res.status(statusCode).json({
-    success: false,
+  // Use statusCode if it's set on the error object
+  if (err.statusCode) {
+    statusCode = err.statusCode;
+  }
+
+  // Use message from error if available
+  if (err.message && err.message !== 'undefined') {
+    message = err.message;
+  }
+
+  // Create standardized error response
+  const errorResponse = createErrorResponse(
     message,
-    ...(process.env.NODE_ENV === "development" && { stack: err.stack }),
-  });
+    code,
+    process.env.NODE_ENV === "development" ? [{ stack: err.stack }] : null,
+    req.id // Request ID if available
+  );
+
+  res.status(statusCode).json(errorResponse);
 };
 
 // Custom error classes
