+++ packages/server/src/middleware/validation/company.js
@@ -1,0 +1,81 @@
+import { body, validationResult } from 'express-validator';
+
+// Validation error handler
+const handleValidationErrors = (req, res, next) => {
+  const errors = validationResult(req);
+  
+  if (!errors.isEmpty()) {
+    return res.status(400).json({
+      success: false,
+      message: 'Validation failed',
+      errors: errors.array().map(error => ({
+        field: error.path,
+        message: error.msg,
+        value: error.value
+      }))
+    });
+  }
+  
+  next();
+};
+
+// Company validation schemas
+export const companySchemas = {
+  create: [
+    body('name')
+      .trim()
+      .isLength({ min: 2, max: 100 })
+      .withMessage('Company name must be between 2 and 100 characters'),
+    body('slug')
+      .optional()
+      .trim()
+      .isLength({ min: 2, max: 50 })
+      .matches(/^[a-z0-9-]+$/)
+      .withMessage('Slug must contain only lowercase letters, numbers, and hyphens'),
+    handleValidationErrors
+  ],
+  update: [
+    body('name')
+      .optional()
+      .trim()
+      .isLength({ min: 2, max: 100 })
+      .withMessage('Company name must be between 2 and 100 characters'),
+    body('slug')
+      .optional()
+      .trim()
+      .isLength({ min: 2, max: 50 })
+      .matches(/^[a-z0-9-]+$/)
+      .withMessage('Slug must contain only lowercase letters, numbers, and hyphens'),
+    handleValidationErrors
+  ]
+};
+
+// Company validation rules
+export const validateCompanyCreation = [
+  body('name')
+    .trim()
+    .isLength({ min: 2, max: 100 })
+    .withMessage('Company name must be between 2 and 100 characters'),
+  body('slug')
+    .optional()
+    .trim()
+    .isLength({ min: 2, max: 50 })
+    .matches(/^[a-z0-9-]+$/)
+    .withMessage('Slug must contain only lowercase letters, numbers, and hyphens'),
+  handleValidationErrors
+];
+
+export const validateCompanyUpdate = [
+  body('name')
+    .optional()
+    .trim()
+    .isLength({ min: 2, max: 100 })
+    .withMessage('Company name must be between 2 and 100 characters'),
+  body('slug')
+    .optional()
+    .trim()
+    .isLength({ min: 2, max: 50 })
+    .matches(/^[a-z0-9-]+$/)
+    .withMessage('Slug must contain only lowercase letters, numbers, and hyphens'),
+  handleValidationErrors
+];
