+++ packages/server/src/routes/authRoutes.js
@@ -17,7 +17,7 @@ const router = express.Router();
 // Public routes
 router.post("/register", validate(authSchemas.register), register);
 router.post("/login", validate(authSchemas.login), login);
-router.post("/google-login", googleLogin); // New Google OAuth endpoint
+router.post("/google-login", validate(authSchemas.googleLogin), googleLogin); // New Google OAuth endpoint
 router.post("/verify-email", validate(authSchemas.verifyEmail), verifyEmail);
 router.post("/resend-verification", validate(authSchemas.resendVerification), resendVerification);
 router.post("/forgot-password", validate(authSchemas.forgotPassword), forgotPassword);
